+++ cmd/netductor-agent/main.go
@@ -19,7 +19,7 @@ import (
 	"time"
 )
 
-var version = "0.8.50"
+var version = "0.8.51"
 
 type config struct {
 	Server   string